requirements.txt
annotated-types==0.7.0
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
docopt==0.6.2
fonttools==4.59.2
Jinja2==3.1.6
MarkupSafe==3.0.2
num2words==0.5.14
numpy==2.3.2
pandas==2.3.2
pdf417==0.8.1
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
-e git+https://github.com/tonicanada/sii_chile_xml_to_pdf.git@cba8c14c89139adb1af69825bf59339899146413#egg=sii_xml_pdf
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
fastapi
uvicorn[standard]
python-multipart
redis
rq
fastapi-mail
Pillow==11.3.0
pytesseract==0.3.13